[435851] Use a message queue to store and render the log messages
Each time a log message is generated, it is added to a message queue. A
task is then created to render the message queue. The task will render
all messages currently available in the queue. This improves
performance, since we could generate log messages faster than we could
render them (at least on MacOS).

Change-Id: I7b38d8ae4fd93a479862952c563cb0f78b77b467
Signed-off-by: Ian Bull <irbull@eclipsesource.com>