Update patch set 4
Patch Set 4:

I have a few concerns with the changes in SetupTaskPerformer.

With respect to this guard:

if ("defaultValue".equals(attribute.getName())


I think this should specifically look for exactly the EAttribute that's the defaultValue attribute of the variable task itself, not for any attribute that happens to have this name, i.e., SetupPackage.Literals.VARIABLE_TASK__DEFAULT_VALUE.  

Also, it's possible (because of escaping) that the fully expanded value can be a string that has content that looks exactly like a variable reference. So the logic with the comment

// if a "defaultValue" attribute is not completely resolved, then nullify it so the user will fill the associated value from clean when prompted 

could find accidental matches. I think the version of expandStrings that takes a keys argument should be used, though for this special case it should be called with an initially empty set and that set should be checked if it's empty after the call.

Patch-set: 4
Label: Code-Review=0