Update patch set 3
Patch Set 3:

I've added it back to the index:

http://git.eclipse.org/c/oomph/org.eclipse.oomph.git/commit/?id=c34f2134d55c9e69d187fdd3143cedf72906614a

I removed the -parent project because it can't resolve unless there is a component.def.  I also added the CBI repo so the license feature can be resolved.  And there are a number of other changes, e.g., user id and password for the mylyn tasks and configurable TP choices.

Patch-set: 3