[506031] Installer hard-codes product branding details
Use the application name and window images configured in the
Eclipse product extension, when available.

Also let a custom installer product tell the framework what is the
installer product ID, so that the SelfUpdate mechanism can correctly
infer installed versions and whether an update is needed, etc.  This
requires a new system property because it is not sufficient just to
introspect the current running product's ID, because we also need to
know whether the current running product is the installer, to change
the run-time behaviour of various components.

https://bugs.eclipse.org/bugs/show_bug.cgi?id=506031
Change-Id: I4fbb1e97e80d13240062856f240113517447a876
Signed-off-by: Christian W. Damus <give.a.damus@gmail.com>