Use UTF-8 workspace encoding in Oomph setup
According to the POMs all sources should be UTF-8. However, the setup
doesn't enforce any encoding, therefore Eclipse takes the platform
default encoding, which at least on Windows is not UTF-8.

The change includes a move of all existing User Preferences, since I
used the preference capture mechanism, which was not yet used before on
this file.

Signed-off-by: Michael Keppler <Michael.Keppler@gmx.de>