Update patch set 3
Patch Set 3:

I prototyped using @none as the value of the oomph.p2.pool system property as a way to specify that you specifically don't want a shared pool and don't want the UI to show such a choice.

Does that still satisfy the original design intent of your contribution?

Patch-set: 3
Label: Verified=0