Update patch set 6
Patch Set 6:

Hi Brian,

I've pushed a new proposal for the management of the OAuth constants. It no longer requires to complicate the build/launch/test process. Instead it obscures the "semi-secret" values directly in the source code. It also adds system properties for all involved values. Please have a look at the new class OAuthConstants and let me know what you think. 90% of the code of that class is reusable across all possible USS applications and it would probably make sense to move it directly into the USS SDK - if you like it...

Please also explain this FIXME tag that you've left. It should be resolved this way or that way.

Patch-set: 6
Label: Verified=0