Update patch set 2
Patch Set 2:

I was able to reproduce the problem by corrupting one of the profiles and deleting profiles.info file in the bundle pool.  It can also be reproduced subsequently by using the Refresh button in the Bundle Pool Management dialog.

I don't believe any change is needed in LazyProfile.snapshot.  Getting the delegate will either return one or throw an exception (when trying to load the profile for the delegate). 

These changes also include improvements to progress monitoring (which was quite poor for my bundle pool with 333 profiles) and to the agent analyzer composite (selecting profiles with > 6000 IUs).

If loading the profile throws an IOException, that is logged as a warning; it's not a problem that can be fixed in the framework but rather a corrupted file on disk, so no point in getting error reports for it. 

Does this all look okay to you?

We can't commit this until after the 1.7 release is wrapped up post Neon.3; hopefully in the coming week sometime.

Patch-set: 2
Reviewer: Gerrit User 1220 <1220@97ee7c02-f12f-4043-b43e-dea463d88b31>
Label: Code-Review=0