Update patch set 2
Patch Set 2:

I've been playing with this and it's really a very nice feature!

I would prefer to move the separate new class PreferenceFilteredTree to be a nested class in PreferenceCaptureDialog.  Is that okay with you?

I've also experimented with further improvements.  In particular the default filtering leaves in place groups, CompoundTasks, with an expansion icon, even when they are empty, so the expansion icon disappears after one tries to expand it, which is normal filter behavior, but not so nice I think.  So I've prototyped an approach where the root nodes too are filtered out as well, which is made tricker by the combination/interaction of the pattern filter and the your default filter.  The result seems even nicer, so I assume you will be happy with such a change.  But I really really don't want to offend you because this is a high quality contribution of significant value, that's why I ask first...

Patch-set: 2