Update patch set 2
Patch Set 2:

Something got messed up again.  I tried to add the change ID and it sort of 1/2 worked stating here that I uploaded a new patch set, but what it shows in the files is not really the commit that I made.  So I cherry picked it onto master and directly committed that:

http://git.eclipse.org/c/oomph/org.eclipse.oomph.git/commit/?id=47feecb99fb715f7f9d954fbf08591aaf9f36695

Patch-set: 2