Update patch set 1
Abandoned

I've tested what the manual import wizard does, and it create an IProject based on the folder name in this case, but that name in the description is ignored because it IProject.create updates the description from what's on disk, so the end result is kind of a strange one.  But if we're going to replicate that with Oomph, we need first to create an IProject handle with the name based on the folder name.

Patch-set: 1
Status: abandoned
Tag: autogenerated:gerrit:abandon