Update patch set 1
Abandoned

Thanks for the inspiring changes and sorry for the slow uptake.

I've committed the following change based on your changes for being able to create a tag:

http://git.eclipse.org/c/oomph/org.eclipse.oomph.git/commit/?id=4e5724653144559ceea47aca17cc41b2abeb6745

This approach is simpler because one doesn't need to change the model.  The checkoutBranch can use a tag name or a branch name and will do the right thing depending on what tags/branches it actually finds in the clone.

Patch-set: 1
Status: abandoned
Tag: autogenerated:gerrit:abandon