Update patch set 1
Patch Set 1: Verified-1 Code-Review-2

This file is generated by org.eclipse.oomph.setup.internal.installer.ProductCatalogGenerator so any changes here will be overwritten. I recall "complaining" about there not being an icon available in the "normal" expected place, but the EPP provider of this package was not interested nor concerned about that. 

In any case, if this is to be fixed it should be in the generator and if the package maintainer doesn't care to provide an icon that fits the pattern of the generator, then it's not clear I should care either.

Patch-set: 1
Label: Verified=-1
Label: Code-Review=-2