Update patch set 1
Patch Set 1: Verified+1 Code-Review+1

This looks good, but I'm holding off until we start the 1.17.0 release cycle for 2020-06 after the 2020-03 release tomorrow.

Thanks for contributing the fix.

It would be great if you could change the commit header to the following to match the style used for all the other commits:

[561102] fixed null child in ContentProvider

We also generally include the link to the Bugzilla in the body of the commit message.

If that's too much hassle, I change it for you.

Patch-set: 1
Label: Verified=+1
Label: Code-Review=+1