Update patch set 1
Patch Set 1:

(1 comment)

> Uploaded patch set 1.

Still not sure about this line:
https://git.eclipse.org/r/plugins/gitiles/oomph/org.eclipse.oomph/+/d5a1a811da32db5ccf54edb14edc5548c1672b84/plugins/org.eclipse.oomph.ui/src/org/eclipse/oomph/internal/ui/FindAndReplaceTarget.java#382

It seems like this label is iternalized in FindReplaceDialog:
https://git.eclipse.org/c/platform/eclipse.platform.text.git/tree/org.eclipse.ui.workbench.texteditor/src/org/eclipse/ui/texteditor/EditorMessages.properties#n106
and this logic will only work for english language. In other cases this label simply will not be replaced and there will remain an analog of the label "Selected lines" for the current locale (different from english one).

If you follow the current logic, you need to have access to the field EditorMessages#FindReplace_SelectedRangeRadioButton_label for proper processing, which I don't know how to organize correctly. So I just left the labels in this place without NON-NLS comments.

Patch-set: 1