Update patch set 2
Patch Set 2:

Sorry, I've been super busy with Eclipse JustJ.  :-(

I did discuss this with Eike and it seems generally okay.  I've not used the platform's externalization approach. I assume it uses quick fixes to create all this stuff but I wanted to play with that to understand it better (because I'll end up having to do this for all future changes).

Also, Eike pointed out there is a PDE preference under Plug-in Development -> Compilers for "Usage of non-externalized strings" that we have set to ignore.  I suppose at least for the projects to be translated this should be set to warning so that it's easier to find all the problems and easier to know when you're not doing that properly...

I'm very sorry that I can't promise to follow up more quickly.  I imagine that's kind of frustrating for you. :-(

Patch-set: 2
Label: Verified=0