#563970: Adding a maven project update setup task
This new task could be used to update maven projects like it could be
done in the m2e UI action "Maven"->"Update Project...".
The task does have 3 parameter, a pattern (0..n) that could be used to
specify a project name pattern which projects should be updated and two
flags ("offline" and "updateSnapshot").

Bug: 563970
Change-Id: I83e6ff7dba58b6f4a26a3966aa68ec2b16802950
Signed-off-by: Martin Schreiber <MARTIN.schreiber@bachmann.info>