Update patch set 1
Patch Set 1: Verified+1 Code-Review+1

Yes the release is out so I should soon be able to proceed.

That being said, there are several incoming reviews in the works (for Oomph and EMF), and I just moved to Switzerland this week, so I'm living in a state of complete chaos.  I've not even set up my workstation yet. :-(

So hopefully I can get my workstation in order very soon and can get clean development environments setup (which is of course easy to do with Oomph setups).  I'll need to reset the version baselines so that changes will come with associated plugin/feature increments.

So please bear with me for a few more days until I come up for air.

Also for this release, I'll need to work on how to consume JustJ JREs...

https://www.eclipse.org/justj/

Patch-set: 1
Label: Verified=+1
Label: Code-Review=+1